docs: add some more notes for troubleshooting (#117)
## Summary

This PR add some error example for troubleshooting and an instruction to
purge minikube (can be useful to restart fresh).

It also replaces the command `kurtosis cluster set docker.k8s` by
`kurtosis cluster set <pick-a-cluster-option>` to force the reader to
actually read the different options and not just copy/paste the command
in the terminal. The options here matters to complete the deployment.