test: decrease the number of nodes spawn (#353)
Decrease the number of Ethereum nodes being spawn when launching the
network locally.

It is not needed to have several nodes to be able to produce blocks only
one EL, CL and validator works fine. This will save some resources.
Unfortunately it doesn't reduce the time to setup by a significant
amount.

---------

Co-authored-by: Steve Degosserie <723552+stiiifff@users.noreply.github.com>